requirements.txt
streamlit==1.43.2
# anthropic[bedrock,vertex]>=0.39.0
jsonschema==4.22.0
# boto3>=1.28.57
google-auth<3,>=2
ruff==0.6.7
pre-commit==3.8.0
pytest==8.3.3
pytest-asyncio==0.23.6
uvicorn[standard]
fastapi
python-dotenv
watchfiles
ruamel.yaml
psutil
httpx
pexpect
